Fix main menu return
Issue: 
The user is not returned to the main menu after tunnel creation is complete. Rather, the user is stuck in the build configuration and forced to hit cancel to return to the main menu. 

Fix: 
Added line `set_state...` in the function `handle_tunnel_creation`.
This makes the user return to the main menu after the tunnel creation is complete. Then the user is given the prompt; `connect...?`.